迁移GoCqHttp的Core实现, 以及封装cq的事件上报为clr事件

+++ SaladimQBot.GoCqHttp/-CqRoot/Posts/-NoticePosts/CqFriendMessageRecalledNoticePost.cs
@@ -6,5 +6,5 @@ public class CqFriendMessageRecalledNoticePost : CqNoticePost
     public Int64 UserId { get; set; }
 
     [Name("message_id")]
-    public Int64 MessageId { get; set; }
+    public Int32 MessageId { get; set; }
 }
